refactor: move format filter functions in respective components
