+++ src/components/OpenFileDialog/VisTypeFilter.js
@@ -5,9 +5,61 @@ import React from 'react'
 import {
     getDisplayNameByVisType,
     visTypeIcons,
+    VIS_TYPE_GROUP_ALL,
+    VIS_TYPE_GROUP_CHARTS,
+    VIS_TYPE_PIVOT_TABLE,
 } from '../../modules/visTypes.js'
 import { VisTypeIcon } from '../VisTypeIcon.js'
 import { CustomSelectOption } from './CustomSelectOption.js'
+
+export const formatTypeFilter = (filterVisTypes, visType) => {
+    const defaultFilterTypes = []
+
+    let defaultTypeFilter
+
+    if (Array.isArray(filterVisTypes)) {
+        defaultFilterTypes.push(
+            ...filterVisTypes
+                .filter(
+                    ({ type, disabled }) =>
+                        !(
+                            disabled ||
+                            [
+                                VIS_TYPE_GROUP_ALL,
+                                VIS_TYPE_GROUP_CHARTS,
+                            ].includes(type)
+                        )
+                )
+                .map(({ type }) => type)
+        )
+
+        if (defaultFilterTypes.length) {
+            defaultTypeFilter = `type:in:[${defaultFilterTypes.join(',')}]`
+        }
+    }
+
+    switch (visType) {
+        case VIS_TYPE_GROUP_ALL: {
+            return defaultTypeFilter
+        }
+        case VIS_TYPE_GROUP_CHARTS: {
+            if (defaultFilterTypes.length) {
+                return `type:in:[${defaultFilterTypes
+                    .filter((item) => item !== VIS_TYPE_PIVOT_TABLE)
+                    .join(',')}]`
+            } else {
+                return `type:!eq:${VIS_TYPE_PIVOT_TABLE}`
+            }
+        }
+        default: {
+            if (visType) {
+                return `type:eq:${visType}`
+            } else if (defaultTypeFilter) {
+                return defaultTypeFilter
+            }
+        }
+    }
+}
 
 export const VisTypeFilter = ({ visTypes, selected, onChange }) => (
     <SingleSelect
